Added nbconvert fix to p4
